requirements.txt
python-telegram-bot==20.7
pandas
yfinance
numpy
nest_asyncio



